+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 Release date `UNRELEASED`
 
 - (chore) Change license to MIT.
+- (test) Fix `lein test` for system's AbstractMethodError.
 
 ## `1.1.0-73`